# Conflicts:
#	src/vm/analysis/read_only_checker/tests.rs

+++ src/chainstate/burn/mod.rs
@@ -61,7 +61,7 @@ pub struct BlockHeaderHash(pub [u8; 32]);
 impl_array_newtype!(BlockHeaderHash, u8, 32);
 impl_array_hexstring_fmt!(BlockHeaderHash);
 impl_byte_array_newtype!(BlockHeaderHash, u8, 32);
-pub const BLOCK_HEADER_HASH_ENCODED_SIZE : u32 = 32;
+pub const BLOCK_HEADER_HASH_ENCODED_SIZE : usize = 32;
 
 pub struct VRFSeed([u8; 32]);
 impl_array_newtype!(VRFSeed, u8, 32);
